Refactor database types into submodules
