+++ lock-keeper/src/types/database.rs
@@ -1,0 +1,4 @@
+//! Models for data stored in the database
+
+pub mod secrets;
+pub mod user;
